// test-api/src/share/enums/commons.enum.ts
export enum TranslationKeys {
  SUCCESSFULLY = 'successfully',
  CREATED = 'created',
  UPDATED = 'updated',
  DELETED = 'deleted',
  INTERNAL_SERVER_ERRORS = 'internal_server_errors',
  ADDED_RULES_COUNT = 'added_rules_count',
  NO_FILE_PROVIDED = 'no_file_provided',
  NO_DATA_FOUND = 'no_data_found',
  IP_INVALID = 'ip_address_format_is_invalid',
  USER_NOT_FOUND = 'user_not_found',
  RULE_EXIST = 'is_already_included_in_rule',
  EMAIL_INVALID = 'invalid_email_address',
  NOT_FOUND = 'not_found',
  RULE_NOT_FOUND = 'rule_not_found',
  UNKNOWN_ERRORS = 'unknown_errors',
  UNAUTHORIZED = 'unauthorized',
  FORBIDDEN = 'forbidden',
  BAD_REQUEST = 'bad_request',
  ACCESS_DENIED = 'access_denied',
  CRON_PROCESSING_RETRY_LATER = 'cron_processing_retry_later',
  ORDER_SCOPE_MISSING = 'order_scope_missing',
  ANALYSIS_IN_PROGRESS = 'analysis_in_progress',
  NEED_PERMISSION_UPDATE_ORDERS = 'need_permission_update_orders',
  PLAN_FEATURE_SUPPORT = 'plan_feature_support',
  DISCOUNT_VALUE_RANGE = 'discount_value_range',
  DISCOUNT_CODE_NOT_FOUND = 'discount_code_not_found',
  CODE_DOES_NOT_EXIST = 'code_does_not_exist',
  CODE_ALREADY_APPLIED = 'code_already_applied',
  APPLYING_CODE_SUCCESSFULLY = 'applying_code_successfully',
  CODE_NOT_ACTIVE = 'code_not_active',
  CODE_NOT_STARTED_YET = 'code_not_started_yet',
  CODE_EXPIRED = 'code_expired',
  SHOP_CANNOT_APPLY = 'shop_cannot_apply',
  LIMIT_USAGE_REACHED = 'limit_usage_reached',
  CREATE_REVIEW_SUCCESSFULLY = 'create_review_successfully',
  LINK_REDIRECT_INVALID = 'link_redirect_invalid',
  RISK_CORE_INVALID = 'risk_core_invalid',
  NO_CHANGES_DETECTED = 'no_changes_detected',
  TITLE_CHAR_LIMIT = 'title_char_limit',
  DESCRIPTION_CHAR_LIMIT = 'description_char_limit',
  DELETED_ITEMS = 'deleted_items',
  PERMISSION_VIEW_ORDERS = 'permission_view_orders',
  SITE_URL_NOT_FOUND = 'site_url_not_found',
  NOTHING_HAS_CHANGED = 'nothing_has_changed',
  AT_LEAST_ONE_RULE = 'at_least_one_rule',
  AT_LEAST_ONE_SETTING_ACTIVE = 'at_least_one_setting_active',
  NOTHING_SET_UP = 'nothing_set_up',
  ORDER_PERMISSION_ALREADY = 'order_permission_already',
  TEMPLATE_NOT_FOUND = 'template_not_found',
  BLOCKING_TEMPLATE_ID_INVALID = 'blocking_template_id_invalid',
  SCOPE_NOT_FOUND = 'scope_not_found',
  PERMISSION_GRANTED = 'permission_granted',
  UPDATED_RULES_TOTAL = 'updated_rules_total',
  NO_VALID_VALUES_PROVIDED = 'no_valid_values_provided',
  PLAN_NOT_FOUND = 'plan_not_found',
  SHOP_NOT_FOUND_OR_PLAN_FREE = 'shop_not_found_or_plan_free',
  CHANGE_PLAN_SUCCESSFULLY = 'change_plan_successfully',
  REQUIRE_READ_CONTENT_PERMISSION = 'require_read_content_permission',
  EVENT_NAME_REQUIRED = 'event_name_required',
  UPDATE_EVENT_SUCCESSFULLY = 'update_event_successfully',
  NOT_AVAILABLE = 'not_available',
  URL_FORMAT_INVALID = 'url_format_invalid',
  REFER_URL_INVALID = 'refer_url_invalid',
  ISP_INVALID = 'isp_invalid',
  ISP_NOT_FOUND = 'isp_not_found',
  RULES_LIMIT_EXCEEDED = 'rules_limit_exceeded',
  INVALID_RULE_FORMAT = 'invalid_rule_format',
  INVALID_COLLECTION = 'invalid_collection',
  COLLECTION_NOT_FOUND = 'collection_not_found',
  PRODUCT_NOT_FOUND = 'product_not_found',
  TIMEZONE_REQUIRED = 'timezone_required',
  PAGE_NOT_FOUND = 'page_not_found',
  REFER_URL_REQUIRED = 'refer_url_required',
  IP_RANGE_FORMAT_INVALID = 'ip_range_format_invalid',
}